+++ gherkin/CHANGELOG.md
@@ -16,6 +16,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 * [Java] Replaced `Gherkin` with a `GherkinParser` that uses a builder to construct.
 * [Java] Made all internal classes package private
 * [Java] Removed unused `stopAtFirstError` property from `Parser`
+* [PHP] Implemented `Gherkin`
 
 ### Deprecated